+++ providers/vpn_jantit.py
@@ -11,10 +11,10 @@ from typing import Optional
 from modules.DB.connectors.mysql import MySql
 from modules.DB.connectors.sqlite import Sqlite3
 from modules.DB.CRUD import CRUD
-from modules.CaptchaSolver import CaptchaSolver
-from modules.Config import Config
-from modules.Logger import Logger
-from modules.Utilities import (write_json_file, read_json_file, get_best_server,
+from modules.captcha_solver import CaptchaSolver
+from modules.config import Config
+from modules.logger import Logger
+from modules.utilities import (write_json_file, read_json_file, get_best_server,
                                generate_random_string, read_config_file, exception_factory)
 ##########################
